% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias_mitigation.R
\name{bias_mitigation}
\alias{bias_mitigation}
\title{Bias mitigation}
\usage{
bias_mitigation(
  method,
  training_data,
  target_variable,
  protected_variable,
  testing_data = NULL,
  probs = NULL,
  lambda = 1,
  cutoff = 0.5,
  features = NULL
)
}
\arguments{
\item{method}{character, Name of the mitigation method
1. "disparate-impact-remover"
2. "reweighing"
3. "resampling"}

\item{training_data}{\code{data.frame}, Data to be transformed}

\item{target_variable}{character, Target variable}

\item{protected_variable}{character, Data column name which contains sensitive information such as gender, race etc...}

\item{testing_data}{\code{data.frame}, Data to be transformed}

\item{probs}{numeric, Vector with probabilities for preferential sampling}

\item{lambda}{numeric, Amount of repair desired for disparate-impact-remover. 
Value from 0 to 1, where 0 will return almost unchanged dataset and 1 fully repaired dataset}

\item{cutoff}{numeric, Threshold for probabilities for sampling.  Value from 0 to 1.}

\item{features}{character, Data columns used to train the model}
}
\value{
Data/weights after the mitigation (list)
}
\description{
Bias mitigation is a wrapper with several options to make a dataset more balanced.
}
\examples{

set.seed(1)
# custom data frame with sex, age and target column names
custom_data <- data.frame(
  sex = c(rep("M", 140), rep("F", 60)),
  age = c(rep(1:20,10)),
  target = c(
  c(rep(c(1, 1, 1, 1, 1, 1, 1, 0, 0, 0),14)),
  c(rep(c(0, 1, 0, 1, 0, 0, 1, 0, 0, 1),6))
  )
)

mitigation_result <- bias_mitigation(
  method = "disparate-impact-remover",
  training_data = custom_data,
  target_variable = "target",
  protected_variable = "sex"
)
}
